refactor & move cmd add logic to scaffolder
